project/huaweiyun_Camera/denoise_project/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0
PySide6>=6.0.0